CLI refactored:  - CLI context routing working  - Somwhat programatic generation of help texts - further work would be    beneficial to make help text, documentation and actual implementation    consistant.  - Global CLI contexts verified with exemptions (1)  - Common CLI contexts verified with exemptions (2)  - Unique MO contexts not verified (3)  (1) CPU and LOG sub-MOs not in place, requires a new work package  including OS support/configuration  (2) Log level per MO/Sub-MO not settled - requires an own work package,  architecture not set.  (3) Needs separate verification with full configuration in place.  The CLI refactoring has generated a set of TRs and improvement ideas  whch have been documented as github issues.
Memory optimizations:
 - Most static- and class- memory allocations are now from external
   SPI-RAM. Optimizations for performance has not yet been performed (4)
 - QLists-, Semaphores- and some libraries does not work with static
   memory assignments (5).

Issues (1) - (5) + many more issues found are now documented in the
github issue list.

A full regression test has not been made (íf we had a release we would
not merge, but at this stage I believe it is the most efficient way to
get to an initial release.

+++ client/Arduino/genJMRIDecoderClient/lgLink.h
@@ -50,6 +50,7 @@
 #include "xmlHelpers.h"
 #include "pinout.h"
 #include "esp32SysConfig.h"
+#include "taskWrapper.h"
 
 class decoder;
 class flash;
@@ -157,10 +158,10 @@ private:
     SemaphoreHandle_t dirtyPixelLock;
     wdt* lgLinkWdt;
     uint32_t overRuns;
-    int64_t maxLatency;
+    uint32_t maxLatency;
     uint32_t maxRuntime;
     uint16_t avgSamples;
-    int64_t* latencyVect;
+    uint32_t* latencyVect;
     uint32_t* runtimeVect;
     uint8_t pin;
     QList<dirtyPixel_t*> dirtyPixelList;
